Move filesystem logic back to cli
