+++ cli/update.go
@@ -44,7 +44,7 @@ func updateRun(cmd *cobra.Command, args []string) {
 	newRequestConfig.Group = group
 	req := controller.NewRequest(newRequestConfig)
 
-	req, err := newController.ExtendWithContent(req)
+	req, err := extendRequestWithContent(fs, req)
 	handleUpdateCmdError(err)
 	req, err = newController.ExtendWithExistingSliceIDs(req)
 	handleUpdateCmdError(err)
