+++ controller/controller.go
@@ -10,8 +10,6 @@ import (
 	"github.com/coreos/fleet/unit"
 
 	"github.com/giantswarm/inago/common"
-	"github.com/giantswarm/inago/file-system/fake"
-	"github.com/giantswarm/inago/file-system/spec"
 	"github.com/giantswarm/inago/fleet"
 	"github.com/giantswarm/inago/task"
 )
@@ -22,8 +20,6 @@ type Config struct {
 	// Dependencies.
 
 	Fleet fleet.Fleet
-
-	FileSystem filesystemspec.FileSystem
 
 	TaskService task.Service
 
@@ -58,7 +54,6 @@ func DefaultConfig() Config {
 
 	newConfig := Config{
 		Fleet:       newFleet,
-		FileSystem:  filesystemfake.NewFileSystem(),
 		TaskService: newTaskService,
 		WaitCount:   3,
 		WaitSleep:   1 * time.Second,
@@ -71,7 +66,6 @@ func DefaultConfig() Config {
 // Controller defines the interface a controller needs to implement to provide
 // operations for groups of unit files against a fleet cluster.
 type Controller interface {
-	ExtendWithContent(req Request) (Request, error)
 	ExtendWithExistingSliceIDs(req Request) (Request, error)
 	ExtendWithRandomSliceIDs(req Request) (Request, error)
